Add support for stdio streams
Extend the API, mimicking `std::process` to provide access to stdin,
stdout and stderr of the subprocess.

To demonstrate basic functionality, add a program doing line-based
copying from stdin to stdout, and drive that using the new API in a
new unit test.

Note that the provided implementation is currently for Unix only, due
to lack of familiarity of the author with the Windows platform.