Implement Read/Write trait in child to support service trait
My intention here is to allow one to write a client service from a
subprocess. I intend to speak with a subprocess on stdin and stdout.

This is probably not the correct way to do it as, stdin and stdout
are hardcoded, but I provided an example to give an idea of what
I would like to write with this crate.

Signed-off-by: Arthur Gautier <baloo@superbaloo.net>