Implement AsRawFd for ChildStd* structs
This allows us to write low-level stuff such as sending child fds over Unix
domain socket. I don't think there was a reason to not expose fd values.

The stub module name "sys" is copied from tokio-tcp.

If PollEvented had implemented AsRawFd, we wouldn't need `.get_ref()`. I'm
not sure if that's a good idea or not, but anyway we'll have to write these
proxy impls.